moved more files to appropriate locations
